refactor: update InternalServeException namespace
