+++ src/Infrastructure/Identity/UserService.CreateUpdate.cs
@@ -1,6 +1,6 @@
-﻿using FSH.WebApi.Application.Common.Exceptions;
+﻿using System.Security.Claims;
+using FSH.WebApi.Application.Common.Exceptions;
 using FSH.WebApi.Application.Common.Mailing;
-using FSH.WebApi.Application.Identity;
 using FSH.WebApi.Application.Identity.Users;
 using FSH.WebApi.Domain.Common;
 using FSH.WebApi.Domain.Identity;
@@ -8,7 +8,6 @@ using FSH.WebApi.Shared.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Identity.Web;
-using System.Security.Claims;
 
 namespace FSH.WebApi.Infrastructure.Identity;
