+++ src/Infrastructure/Identity/UserService.Password.cs
@@ -1,6 +1,5 @@
 ﻿using FSH.WebApi.Application.Common.Exceptions;
 using FSH.WebApi.Application.Common.Mailing;
-using FSH.WebApi.Application.Identity;
 using FSH.WebApi.Application.Identity.Users.Password;
 using Microsoft.AspNetCore.WebUtilities;
